Set the fixed start header HASH to get block headers info
Set the correctly block header ID to query from other peers,
Now the node could receive the block headers correctly

+++ net/node.go
@@ -133,7 +133,7 @@ func (node *node) rx() error {
 
 		switch err {
 		case nil:
-			go handleNodeMsg(node, buf, len)
+			go handleNodeMsg(node, buf, uint64(len))
 			break
 		case io.EOF:
 			//log.Println("Reading EOF of network conn")
